Add endpoint for getting list of components with their versions (#324)
1. Changed the registry to anticipate 'Accept: application/vnd.oc.info+json'
HTTP header and if detect it to skip the component rendering and to return
just the component info.

2. Added a new method in Client - getComponentsInfo.
It prepares the request to the registry with the proper accept header, checks
for errors and provides a processed pair (error, info).
- in case info data is provided for all the requested components then the error
will be null and the info will have the following structure:
[{
   componentName: '<name>',
   requestedVersion: '<requested version|undefined>'
   apiResponse: {  //// comes from the registry
      href: '<component url>',
      name: '<name>',
      renderMode: '<rendered|unrendered>',
      requestVersion: <requested version>,
      type: '<type>',
      version: '<actual version>'
   }
},
{
   componentName: '<name>',
   requestedVersion: '<requested version|undefined>'
   apiResponse: {...}
},
...
]

- in case of errors then the info part will skip the apiResponse and will
include an error field:
{
   componentName: '<name>',
   requestedVersion: '<requested version|undefined>'
   error: ...
}

In addition to this the 'error' parameter in the callback will be an array with
the error element corresponding to each component requested. If there's no
error for a particular component then the corresponding element in the 'error'
array will be null:
[
   null,  //// means the first component is OK
   { message: ..., stackTrace: ... },
   null,  //// third components is OK as well
   ...
]