Fix for Custom eaders of one component can appear to another.
The problem was introduced with #326 -
In get-component.js the declaration of responseHeaders variable happens in a
more global scope than it should be. For this reason every invocation of
renderer function sees the changes in the headers made by previously requested
components.
The fix is to just move the declaration one scope inner.

- Added a unit test to validate the fix - it tries to load two components
subsequently. The first one does provide custom headers while the second
one - don't. Then we check and expect to see there are no custom headers
in the result for the second one.