333 - Implement customHeadersToSkipOnWeakVersion functionality
- Introduced new registry option - customHeadersToSkipOnWekVersion which expects
to be an array of strings denoting the headers that have to be omitted from
the response in case a weak version is requested (i.e 1.x.x). The rational
behind this is the need to control cache related headers - in case 1.x.x
version is requested we would not like cache headers to be returned because
otherwise the response will be cached. Next time when someone asks for 1.x.x
it will get the cached response even though a new version may be available.
- added unit and integration tests